GH-95 Initial refactoring and code cleanup
Moved private classes Node and RangeIterator and interface NodeListener out of BTree.
The classes are package-private now.
Changed visibility of some fields and methods of BTree to package-private for that.
Moved the test/diagnostics code to tests.
Removed the commented-out code.
